Remove committee member is now a separate modal. Committee heads can remove members from the committee overview page. (TT-42)
